rename fetchWithHttpCache
